dom/fetch/tests/test_fetch_system.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Test for fetch() with system privileges</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  </head>
  <body onload="runTests();">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test">
<script class="testbody" type="application/javascript">
let tests = [];

// Re-using the resources from the similar XHR tests.
const PROTECTED_URL = "file:///etc/passwd";
const REDIRECT_URL =
  window.location.protocol +
  "//" +
  window.location.host +
  "/tests/dom/xhr/tests/file_XHR_system_redirect.html";
const CROSSSITE_URL =
  "http://example.com/tests/dom/xhr/tests/test_XHR_system.html";

tests.push(async () => {
  // fetch with systemXHR permission can load cross-origin resources.

  is(window.location.hostname, "mochi.test", "correct origin");

  try {
    let response = await window.fetch(CROSSSITE_URL, { mode: "no-cors" });
    is(response.status, 200, "correct HTTP status");
    let text = await response.text();
    ok(text != null, "HTTP response non-null");
    ok(text.length, "HTTP response not empty");
    runNextTest();
  } catch (e) {
    ok(false, "Got a fetch error: " + e);
    runNextTest();
  }
});

tests.push(async () => {
  // fetch with systemXHR permission is not permitted to access file:/// URIs.

  try {
    let response = await window.fetch(PROTECTED_URL, { mode: "no-cors" });
    ok(false, "Should not have loaded");
    runNextTest();
  } catch (e) {
    ok(true, "Got a fetch error: " + e);
    runNextTest();
  }
});

tests.push(async () => {
  // fetch with systemXHR permission won't load file:/// URIs even if an HTTP resource redirects there.

  try {
    let response = await window.fetch(REDIRECT_URL, { mode: "no-cors" });
    ok(false, "Should not have loaded");
    runNextTest();
  } catch (e) {
    ok(true, "Got a fetch error: " + e);
    runNextTest();
  }
});

function runNextTest() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }
  tests.shift()();
}

function runTests() {
  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPermissions(
    [{ type: "systemXHR", allow: true, context: document }],
    runNextTest
  );
}
</script>
</pre>
  </body>
</html>
